+++ app/Http/Controllers/RatingController.php
@@ -1,6 +1,8 @@
 <?php
 
 namespace App\Http\Controllers;
+
+use App\Models\AdminNotification;
 use App\Models\Category;
 use App\Models\Product;
 use App\Models\Rating;
@@ -16,6 +18,9 @@ use Illuminate\Support\Facades\Validator;
 class RatingController extends Controller
 {
     public static function submitReview(Request $request){
+        if(Auth::guard('admin')->user()->access->review_manage != 3){
+            return redirect(route('admin.dashboard'))->with('error', 'Access Denied!');
+        }
         $validator = Validator::make($request->all(),[
             'product_id' => 'required',
             'name' => 'required',
@@ -33,6 +38,7 @@ class RatingController extends Controller
               'rating' =>  $request->rating,
             ];
             Notification::send($admin, new NewReviewNotfication($data));
+            
             return back();
         }else{
             return back()->withErrors($validator);
@@ -40,7 +46,9 @@ class RatingController extends Controller
     }
 
     public static function reviews(){
-
+        if(Auth::guard('admin')->user()->access->review_manage == 2){
+            return redirect(route('admin.dashboard'))->with('error', 'Access Denied! You do not have permission to access this page.');
+        }
         return view('admin.reviews.reviews',[
             'admin' => Auth::guard('admin')->user(),
             'siteSettings' => SiteSetting::where('id', 1)->first(),
@@ -49,17 +57,36 @@ class RatingController extends Controller
     }
 
     public static function reviewDestroy(Request $request, $id){
+        if(Auth::guard('admin')->user()->access->review_manage != 3){
+            return redirect(route('admin.dashboard'))->with('error', 'Access Denied!');
+        }
         $rating = Rating::find($id);
         if (isset($rating)){
+            $notification = new AdminNotification();
+                $notification->admin_id = auth()->id();
+                $notification->message = $rating->product->name .' Review has been deleted.';
+                $notification->notification_for = 'Review';
+                $notification->item_id = $rating->id;
+                $notification->save();
+
             $rating->delete();
             $successMessage = "Your Review has been deleted";
             $request->session()->flash('success', $successMessage);
+
+            
+
             return back();
         }
     }
 
     public static function reviewShow($id){
-        Rating::statusCheck($id);
+        $rating = Rating::statusCheck($id);
+        $notification = new AdminNotification();
+                $notification->admin_id = auth()->id();
+                $notification->message = $rating->product->name .' Review status has been updated.';
+                $notification->notification_for = 'Review';
+                $notification->item_id = $rating->id;
+                $notification->save();
         return back();
     }
 }
